+++ docs/conf.py
@@ -20,7 +20,6 @@ extensions = [
     "sphinx_design",
     "sphinx_reredirects",
     "sphinx.ext.intersphinx",
-    "sphinx.ext.autosectionlabel",
     "sphinx.ext.napoleon",
     "sphinx_autodoc_typehints",  # must be loaded after sphinx.ext.napoleon. See https://github.com/tox-dev/sphinx-autodoc-typehints/issues/15
 ]
@@ -115,8 +114,14 @@ html_static_path = [
 html_js_files = [
     ("js/custom-icons.js", {"defer": "defer"}),
 ]
-napoleon_use_param = True
 pygments_style = "sphinx"
+
+
+# -- Napolean configuration ----------------------------------
+napoleon_use_param = True
+napoleon_google_docstring = True
+napoleon_attr_annotations = True
+
 
 # -- Intersphinx configuration ----------------------------------