+++ src/icalendar/cal/alarm.py
@@ -295,7 +295,7 @@ class Alarm(Component):
             :class:`Alarm`
 
         Raises:
-            InvalidCalendar: If the content is not valid according to :rfc:`5545`.
+            ~error.InvalidCalendar: If the content is not valid according to :rfc:`5545`.
 
         .. warning:: As time progresses, we will be stricter with the validation.
         """