+++ src/icalendar/cal/free_busy.py
@@ -130,7 +130,7 @@ class FreeBusy(Component):
             :class:`FreeBusy`
 
         Raises:
-            InvalidCalendar: If the content is not valid according to :rfc:`5545`.
+            ~error.InvalidCalendar: If the content is not valid according to :rfc:`5545`.
 
         .. warning:: As time progresses, we will be stricter with the validation.
         """